requirements.txt
python-dotenv
google-generativeai
langchain
langchain-google-genai
langchain_community
chromadb
gradio